# swagger.yml
swagger: "2.0"
info:
  version: "1.0.0"
  title: "Sample API"
  description: "API documentation for a data flow sync"
schemes:
  - http
  - https
security:
  - api_key: []
basePath: /api
consumes: [application/json]
produces: [application/json]
...
paths:
...
  /user/verify-account:
    post:
      summary: verify signed up account
      operationId: VerifyAccount
      tags:
        - "account"
      security: 
        - api_key: []
      parameters:
        - in: body
          name: Account-verify
          required: true
          schema:
            type: object
            properties: 
              otp: 
                type: string
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/Token"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          $ref: "#/responses/Default"
  /user/send-mail:
    post:
      summary: send otp to specified email address
      operationId: SendMail
      tags:
        - "account"
      security: 
        - api_key: []
      responses:
        200:
          description: OK
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
        500:
          description: Internal Error
          schema:
            $ref: "#/definitions/ErrorResponse"
        default:
          $ref: "#/responses/Default"

  # limitary zone
...
securityDefinitions:
  api_key:
    type: apiKey
    in: header
    name: x-data-flow-sync-auth-key
responses:
  Default:
    description: Unexpected error
    schema:
      $ref: "#/definitions/ErrorResponse"
definitions:
...
  ErrorResponse:
    type: object
    properties:
      code:
        type: integer
        description: error code
      details:
        type: string
        description: error details
  Token:
    type: object
    properties:
      token:
        type: string
...